refactor: use int64 for ids

+++ internal/infra/sqlite/sqlite_episodes.go
@@ -24,11 +24,11 @@ import (
 
 func (s Repository) GetEpisode(
 	ctx context.Context,
-	userid, podcastid int32,
+	userid, podcastid int64,
 	episode string,
 ) (*model.Episode, error) {
 	logger := log.Ctx(ctx)
-	logger.Debug().Int32("user_id", userid).Int32("podcast_id", podcastid).Str("episode", episode).
+	logger.Debug().Int64("user_id", userid).Int64("podcast_id", podcastid).Str("episode", episode).
 		Msgf("get episode")
 
 	query := `
@@ -59,13 +59,13 @@ func (s Repository) GetEpisode(
 // When aggregate get only last action for each episode.
 func (s Repository) ListEpisodeActions(
 	ctx context.Context,
-	userid int32, deviceid, podcastid *int32,
+	userid int64, deviceid, podcastid *int64,
 	since time.Time,
 	aggregated bool,
 	lastelements uint,
 ) ([]model.Episode, error) {
 	logger := log.Ctx(ctx)
-	logger.Debug().Int32("user_id", userid).Any("podcast_id", podcastid).Any("device_id", deviceid).
+	logger.Debug().Int64("user_id", userid).Any("podcast_id", podcastid).Any("device_id", deviceid).
 		Msgf("get episodes since=%s aggregated=%v", since, aggregated)
 
 	query := `
@@ -124,9 +124,9 @@ func (s Repository) ListEpisodeActions(
 	return episodesFromDb(res), nil
 }
 
-func (s Repository) ListFavorites(ctx context.Context, userid int32) ([]model.Episode, error) {
+func (s Repository) ListFavorites(ctx context.Context, userid int64) ([]model.Episode, error) {
 	logger := log.Ctx(ctx)
-	logger.Debug().Int32("user_id", userid).Msg("get favorites")
+	logger.Debug().Int64("user_id", userid).Msg("get favorites")
 
 	query := `
 		SELECT e.id, e.podcast_id, e.url, e.title, e.guid, e.created_at, e.updated_at,
@@ -149,10 +149,10 @@ func (s Repository) ListFavorites(ctx context.Context, userid int32) ([]model.Ep
 }
 
 func (s Repository) GetLastEpisodeAction(ctx context.Context,
-	userid, podcastid int32, excludeDelete bool,
+	userid, podcastid int64, excludeDelete bool,
 ) (*model.Episode, error) {
 	logger := log.Ctx(ctx)
-	logger.Debug().Int32("user_id", userid).Int32("podcast_id", podcastid).
+	logger.Debug().Int64("user_id", userid).Int64("podcast_id", podcastid).
 		Msgf("get last episode action excludeDelete=%v", excludeDelete)
 
 	query := `
@@ -187,9 +187,9 @@ func (s Repository) GetLastEpisodeAction(ctx context.Context,
 	return res.toModel(), nil
 }
 
-func (s Repository) SaveEpisode(ctx context.Context, userid int32, episodes ...model.Episode) error {
+func (s Repository) SaveEpisode(ctx context.Context, userid int64, episodes ...model.Episode) error {
 	logger := log.Ctx(ctx)
-	logger.Debug().Int32("user_id", userid).Msg("save episode")
+	logger.Debug().Int64("user_id", userid).Msg("save episode")
 
 	dbctx := db.MustCtx(ctx)
 
@@ -207,10 +207,10 @@ func (s Repository) SaveEpisode(ctx context.Context, userid int32, episodes ...m
 	for _, episode := range episodes {
 		logger.Debug().Object("episode", &episode).Msg("save episode")
 
-		deviceid := sql.NullInt32{}
+		deviceid := sql.NullInt64{}
 		if episode.Device != nil {
 			deviceid.Valid = true
-			deviceid.Int32 = episode.Device.ID
+			deviceid.Int64 = episode.Device.ID
 		}
 
 		if episode.Timestamp.IsZero() {
