+++ internal/service/subscriptions.go
@@ -261,7 +261,7 @@ func (s *SubscriptionsSrv) getUser(ctx context.Context, username string) (*model
 // getUserDevice return device from database and mark last_seen.
 func (s *SubscriptionsSrv) getUserDevice(
 	ctx context.Context,
-	userid int32,
+	userid int64,
 	devicename string,
 ) (*model.Device, error) {
 	device, err := s.devicesRepo.GetDevice(ctx, userid, devicename)
